% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groundwater}
\alias{groundwater}
\title{Groundwater data}
\format{A data.frame with 10 variables: \code{year}, \code{site},
\code{catch}, \code{pos}ition, \code{month}, \code{no3mgl}, \code{no3mgladj},
\code{po4mgl}, and \code{watershed}.}
\usage{
groundwater
}
\description{
Groundwater data for monthly nitrate and phosphate.
}
\seealso{
\code{\link{surfacewater}}
}
\keyword{datasets}

