% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surfacewater}
\alias{surfacewater}
\title{Surface water data}
\format{A data.frame with 10 variables: \code{year}, \code{site},
\code{catch}, \code{runoff_mm}, \code{sed_kgha}, \code{tn_kgha},
\code{tp_kdha}, \code{no3_kgha}, \code{sed_mgL}, \code{tn_mgL},
\code{tp_mgL}, \code{no3_mgL}, and \code{watershed}.}
\usage{
surfacewater
}
\description{
Surface water data for monthly nitrate and phosphate.
}
\examples{
# Calculate average all crop sediment, total nitrogen, and phosphorus loads.
library(dplyr)
surfacewater \%>\%
  left_join(STRIPSMeta::watersheds) \%>\%
  filter(prairie_pct == 0, year>2007) \%>\%
  select(watershed, sed_kgha, tn_kgha, tp_kgha) \%>\%
  summarize(sed_Mgha = mean(sed_kgha, na.rm = TRUE)/1000,
            tn_kgha  = mean(tn_kgha,  na.rm = TRUE),
            tp_kgha  = mean(tp_kgha,  na.rm = TRUE))
}
\seealso{
\code{\link{groundwater}}
}
\keyword{datasets}
