% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surfacewater}
\alias{surfacewater}
\title{Surface water data}
\format{A data.frame with 10 variables: \code{year}, \code{site},
\code{catch}, \code{runoff_mm}, \code{sed_kgha}, \code{tn_kgha},
\code{tp_kdha}, \code{no3_kgha}, \code{sed_mgL}, \code{tn_mgL},
\code{tp_mgL}, \code{no3_mgL}, and \code{watershed}.}
\usage{
surfacewater
}
\description{
Surface water data for monthly nitrate and phosphate.
}
\seealso{
\code{\link{groundwater}}
}
\keyword{datasets}

